+++ backend/common/watchdogs/group_llms.py
@@ -10,20 +10,22 @@ class GroupLLMsTable:
 
     @classmethod
     async def async_select_by_group_id_order_by_seqn(cls, session: AsyncSession, group_id: int) -> list[GroupLLMs]:
-        where_clause = (GroupLLMs.deleted == 0) & (GroupLLMs.group_id == group_id)
+        where_clause = (GroupLLMs.deleted == 0) & (GroupLLMs.enabled == True) & (GroupLLMs.group_id == group_id)
         stmt = select(GroupLLMs).where(where_clause).order_by(GroupLLMs.gllms_seqn)
         result = await session.execute(stmt)
         return list(result.scalars().all())
 
     @classmethod
     async def async_select_all_order_by_group_id_seqn(cls, session: AsyncSession) -> list[GroupLLMs]:
-        stmt = select(GroupLLMs).where(GroupLLMs.deleted==0).order_by(GroupLLMs.group_id, GroupLLMs.gllms_seqn)
+        where_clause = (GroupLLMs.deleted == 0) & (GroupLLMs.enabled == True)
+        stmt = select(GroupLLMs).where(where_clause).order_by(GroupLLMs.group_id, GroupLLMs.gllms_seqn)
         result = await session.execute(stmt)
         return list(result.scalars().all())
     
     @classmethod
-    def sync_select_all(cls, session: Session) -> list[GroupLLMs]:
-        stmt = select(GroupLLMs).where(GroupLLMs.deleted==0).order_by(GroupLLMs.gllms_id)
+    def sync_select_all_enabled(cls, session: Session) -> list[GroupLLMs]:
+        where_clause = (GroupLLMs.deleted == 0) & (GroupLLMs.enabled == True)
+        stmt = select(GroupLLMs).where(where_clause).order_by(GroupLLMs.gllms_id)
         return list(session.execute(stmt).scalars().all())
     
     @classmethod